# api.yml
...
swagger: '2.0'
info:
  version: 1.0.0
  title: Microservice for Books
  description: Java + MySQL
  license:
    name: Software Architecture 2018-II
host: 192.168.99.101:8000
basePath: /books-ms
schemes:
  - http
consumes:
  - application/json
  - multipart/form-data
produces:
  - application/json
paths:
...
  /books/{book_id}/genre/{genre_id}:
    post:
      summary: Add a genre to a book
      operationId: addGenreToBook
      parameters:
        - name: book_id
          in: path
          required: true
          description: The id of the book
          type: number
        - name: genre_id
          in: path
          required: true
          description: The id of the genre
          type: number
      responses:
        '201':
          description: Created
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
    
    delete:
      summary: Remove a genre of a book
      operationId: deleteGenreOfBook
      parameters:
        - name: book_id
          in: path
          description: The id of the book
          required: true
          type: number
        - name: genre_id
          in: path
          required: true
          description: The id of the genre
          type: number
      responses:
        '200':
          description: Ok
        '404':
          description: Not Found
        '500':
          description: Internal Server Error'
...
  /books/{book_id}/genres:
    get:
      summary: Returns genres a book with given id
      operationId: findGenresOfBook
      parameters:
        - name: book_id
          in: path
          description: The title of the book to be returned
          required: true
          type: number
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/GenresList'
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
...
  /genres:
    post:
      summary: Creates a new genre
      operationId: createGenre
      parameters:
        - name: genre
          in: body
          required: true
          description: Genre Object
          schema:
            $ref: '#/definitions/GenreInput'
      responses:
        '201':
          description: Created
        '500':
          description: Internal Server Error
    get:
      summary: Returns all genres
      operationId: findAllGenres
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/GenresList'
        '500':
          description: Internal Server Error
  /genres/{genre_id}:
    get:
      summary: Returns a genre with the given id
      operationId: findByGenreId
      parameters:
        - name: genre_id
          in: path
          description: The id of the genre to be returned
          required: true
          type: number
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/Genre'
        '404':
          description: Not Found
        '500':
          description: Internal Server Error

    put:
      summary: Updates a genre with the given id
      operationId: updateGenre
      parameters:
        - name: genre_id
          in: path
          description: The id of the Genre to be updated
          required: true
          type: number
        - name: Genre
          in: body
          required: true
          description: Genre Object
          schema:
            $ref: '#/definitions/GenreInput'
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
        '500':
          description: Internal Server Error

    delete:
      summary: Deletes a genre with the given id
      operationId: deleteGenre
      parameters:
        - name: genre_id
          in: path
          description: The id of the genre to be deleted
          required: true
          type: number
      responses:
        '200':
          description: Ok
        '404':
          description: Not Found
        '409':
            description: Conflict
        '500':
          description: Internal Server Error
  /genres/name/{genre_name}:
    get:
      summary: Returns genres with the given name
      operationId: findByGenreName
      parameters:
        - name: genre_name
          in: path
          description: The name of the genres to be returned
          required: true
          type: string
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/GenresList'
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /genres/{genre_id}/books:
    get:
      summary: Returns books of a genre with the given id
      operationId: findBooksOfGenre
      parameters:
        - name: genre_id
          in: path
          description: The id of the genre
          required: true
          type: number
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/BooksList'
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /file/book/{book_id}:
    post: 
      summary: Uploads cover to a book with given id 
      operationId: uploadBookCover
      parameters:
        - name: book_id
          in: path
          description: The id of the book
          required: true
          type: number
        - name: file
          in: formData
          description: cover to upload
          required: true
          type: file
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/File'
        '404':
          description: Not Found
        '409':
          description: Conflict
        '500':
          description: Internal Server Error
  
    put: 
      summary: Updates cover of a book with given id 
      operationId: updateBookCover
      parameters:
        - name: book_id
          in: path
          description: The id of the book
          required: true
          type: number
        - name: file
          in: formData
          description: cover to upload
          required: true
          type: file
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/File'
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  
    delete: 
      summary: Deletes cover of a book with given id 
      operationId: deleteBookCover
      parameters:
        - name: book_id
          in: path
          description: The id of the book
          required: true
          type: number
      responses:
        '200':
          description: Ok
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /file/author/{author_id}:
    post: 
      summary: Uploads photo to a author with given id 
      operationId: uploadAuthorPhoto
      parameters:
        - name: author_id
          in: path
          description: The id of the author
          required: true
          type: number
        - name: file
          in: formData
          description: photo to upload
          required: true
          type: file
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/File'
        '404':
          description: Not Found
        '409':
          description: Conflict
        '500':
          description: Internal Server Error
  
    put: 
      summary: Updates photo of a author with given id 
      operationId: updateAuthorPhoto
      parameters:
        - name: author_id
          in: path
          description: The id of the author
          required: true
          type: number
        - name: file
          in: formData
          description: photo to upload
          required: true
          type: file
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/File'
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  
    delete: 
      summary: Deletes photo of a author with given id 
      operationId: deleteAuthorPhoto
      parameters:
        - name: author_id
          in: path
          description: The id of the author
          required: true
          type: number
      responses:
        '200':
          description: Ok
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
definitions:
...
  File:
    type: object
    properties:
      fileName:
        type: string
      fileType:
        type: string
      fileDownloadUri:
        type: string
      size:
        type: number  
  
  # Books
  Book:
    type: object
    properties:
      id:
        type: number
      publisher:
        type: string
      title:
        type: string
      numPages:
        type: number
      isbn:
        type: string
      plot:
        type: string
      cover:
        $ref: '#/definitions/File'
...
  BooksList:
    type: object
    required:
      - list
    properties:
      list:
        type: array
        items:
          $ref: '#/definitions/Book'

  # Author
...
  Genre:
    type: object
    properties:
      id:
        type: number
      name:
        type: string
      description:
        type: string
      photo:
        $ref: '#/definitions/File'
  GenreInput:
    type: object
    properties:
      name:
        type: string
      description:
        type: string
  GenresList:
    type: object
    required:
      - list
    properties:
      list:
        type: array
        items:
          $ref: '#/definitions/Genre'